==> Analysis_Code/7_replication_met_mesa.Rmd <==
---
title: "Metabolomic Risk Score Replication in MESA"
author: "Iain R. Konigsberg"
date: '`r Sys.Date()`'
output: 
  rmarkdown::html_document:
    code_folding: show
    toc: true
    toc_float: 
      collapsed: false
    df_print: paged
    number_sections: true
---

# Background

This Rmd contains code to apply metabolomic risk score models to MESA participants. 

# Set Up

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = "C:/Users/konigsbi/Desktop/ors_topmed/")
```



```{r message=F}
library(ggplot2)
library(glmnetUtils)
library(data.table)
library(readxl)
library(stringr)
library(sjPlot)

setwd("C:/Users/konigsbi/Desktop/ors_topmed/")

start = Sys.time()

theme_set(theme_bw())
set.seed(11)

replicate = function(omics, model, lambda = "lambda.min"){
  weights = as.data.frame(as.matrix(coef(model, s = lambda)))
  weights$dummy = NA
  weights = weights[weights$s1 != 0, ]
  print(paste0("Number of Non-Zero model weights: ", nrow(weights) - 1))
  print(paste0("Number of Non-Zero model weights in replication sample: ", nrow(weights[rownames(weights) %in% colnames(omics), ])))
  
  inty = weights$s1[1]
  weights = weights[2:nrow(weights), ]
  
  weights = weights[rownames(weights) %in% colnames(omics), ]
  omics_sub = omics[, colnames(omics) %in% rownames(weights)] 
  
  omics_sub = omics_sub[, !duplicated(colnames(omics_sub))]
  
  weights = weights[order(rownames(weights)), ]
  omics_sub = omics_sub[, order(colnames(omics_sub))]
  
  print(paste0("Model Weights & Omics Line Up: ", all(rownames(weights) == colnames(omics_sub))))
  omics_sub = t(omics_sub)
  
  scores = as.data.frame(t(weights$s1 %*% as.matrix(omics_sub) + inty))
  names(scores) = "score"
  return(scores)
}

model_select = function(mod){
  index = 1
  df = NULL
  for (x in 1:11){
    alpha = mod$alpha[x]
    moddy = mod$modlist[[x]]
    for (y in 1:length(moddy$lambda)){
      lambda = moddy$lambda[y]
      nonzero = moddy$nzero[y]
      measure = moddy$cvm[y]
      se = moddy$cvsd[y]
      df = rbind(df, c(alpha, index, lambda, nonzero, measure, se))
    }
    index = index + 1
  }
  df = as.data.frame(df)
  names(df) = c("alpha", "index", "lambda", "nonzero", "measure", "se")
  
  df = df[df$nonzero != 0, ]
  
  min_measure = df[df$measure == min(df$measure), ]$measure
  se_measure = df[df$measure == min(df$measure), ]$se
  sub_df = df[df$measure <= min_measure + se_measure, , drop = F]
  sub_df = sub_df[sub_df$alpha == max(sub_df$alpha), , drop = F]
  alpha_elastic = sub_df[sub_df$lambda == max(sub_df$lambda), ]$alpha
  lambda_elastic = sub_df[sub_df$lambda == max(sub_df$lambda), ]$lambda
  nonzero_elastic = sub_df[sub_df$lambda == max(sub_df$lambda), ]$nonzero
  alpha_index = sub_df[sub_df$lambda == max(sub_df$lambda), ]$index
  return(c(alpha_index, lambda_elastic, alpha_elastic))
}


```

# Prepare Metabolites

```{r}
pd = fread("Data/MESA/SHARe Phenotype Datasets/MESAExam1_5/SHARe_Exam1Main.txt") #6431
pd = as.data.frame(pd) %>% dplyr::select(sidno, age1c, race1c, gender1, site1c, bmi1c, pkyrs1c, cig1c, htcm1)
pd2 = fread("Data/MESA/SHARe Phenotype Datasets/MESAExam1_5/SHARe_Exam5Main.txt")
pd2 = as.data.frame(pd2) %>% dplyr::select(sidno, age5c, site5c, bmi5c, pkyrs5c, cig5c, htcm5)
pd = merge(pd, pd2, by = "sidno", all = T)
#should take note of exclusions
spiro = as.data.frame(fread("Data/MESA/SHARe Phenotype Datasets/Lung/SHARe_AncilMesaLungSpirometryExam5.txt")) 
spiro = spiro %>% dplyr::select(sidno, fev1ab5, fvcab5, fevfvcab5, cpgoldab5, cpstgab5, f2575ab5, pefab5)
pd = merge(pd, spiro, by = "sidno", all = T)
ct = as.data.frame(fread("Data/MESA/SHARe Phenotype Datasets/Lung/SHARe_AncilMesaLungExam1CT.txt")) #6425. 3853 w/ exam1 scans.  
ct = ct %>% dplyr::select(sidno, B_tot_vx, B_be_950, model) 
ct$pctemph = ct$B_be_950/ct$B_tot_vx #needs to be log transformed for analysis. 
pd = merge(pd, ct, by = "sidno", all = T)
airway = as.data.frame(fread("Data/MESA/SHARe Phenotype Datasets/Lung/SHARe_AncilMesaLungAirway.txt"))
airway = airway %>% dplyr::select(sidno, wall_thickness, wa_pct)
pd = merge(pd, airway, by = "sidno", all = T)

met = as.data.frame(read_excel("Data/MESA/Metabolomics/Copy of Released Clean TopMed-Post-normalized-data-20170316.xlsx")) 

namey = as.data.frame(readxl::read_excel("Data/MESA/MESA_TOPMed_WideID_20190517.xlsx", skip = 2))
namey = namey %>% dplyr::select(sidno, tom_id1, tom_id5)
namey_one = namey[!is.na(namey$tom_id1), ]
namey_one = namey_one[namey_one$tom_id1 %in% met$TOM_ID, ] 
namey_five = namey[!is.na(namey$tom_id5), ]
namey_five = namey_five[namey_five$tom_id5 %in% met$TOM_ID, ] 

met_one = met[met$TOM_ID %in% namey$tom_id1, ] 
met_one = met_one[!is.na(met_one$TOM_ID), ] 
met_five = met[met$TOM_ID %in% namey$tom_id5, ] 
met_five = met_five[!is.na(met_five$TOM_ID), ] 

rownames(met_one) = met_one$TOM_ID
met_one = met_one[, 6:ncol(met_one)] 
rownames(met_five) = met_five$TOM_ID
met_five = met_five[, 6:ncol(met_five)] 

met1 = as.data.frame(read_excel("Data/MESA/Metabolomics/MESA_pilot_BroadInst_C8-pos_lipids_050517.xlsx", skip = 4)) 
met2 = as.data.frame(read_excel("Data/MESA/Metabolomics/MESA_pilot_BroadInst_HIL-pos_polar_050517.xlsx", skip = 4)) 

met1 = met1[1:321, ]
met1$Metabolite = make.unique(met1$Metabolite)
rownames(met1) = met1$Metabolite 
met1 = met1[, 7:ncol(met1)]
met1 = as.data.frame(t(met1))

met2 = met2[1:253, ]
met2$Metabolite = make.unique(met2$Metabolite)
rownames(met2) = met2$Metabolite
met2 = met2[, 7:ncol(met2)]
met2 = as.data.frame(t(met2))

met_one = merge(met_one, met1, by = "row.names")
dim(met_one) 
rownames(met_one) = met_one$Row.names
met_one$Row.names = NULL
met_one = merge(met_one, met2, by = "row.names")
dim(met_one) 
rownames(met_one) = met_one$Row.names
met_one$Row.names = NULL

met_five = merge(met_five, met1, by = "row.names")
dim(met_five) 
rownames(met_five) = met_five$Row.names
met_five$Row.names = NULL
met_five = merge(met_five, met2, by = "row.names")
dim(met_five) 
rownames(met_five) = met_five$Row.names
met_five$Row.names = NULL

for (x in 1:nrow(met_one)){
  rownames(met_one)[x] = met[met$TOM_ID == rownames(met_one)[x], ]$sidno
}
for (x in 1:nrow(met_five)){
  rownames(met_five)[x] = met[met$TOM_ID == rownames(met_five)[x], ]$sidno
}


conv = read.csv("Data/MESA/Metabolomics/COPDGene_MESA_MetaboliteMapping_edited.csv")

#replace feature ids with metabo ids when possible. 
for (x in 1:ncol(met_one)){
  if (colnames(met_one)[x] %in% conv$name_mesa){
    colnames(met_one)[x] = na.omit(conv[conv$name_mesa == colnames(met_one)[x], ]$metabid)
  }
}
for (x in 1:ncol(met_five)){
  if (colnames(met_five)[x] %in% conv$name_mesa){
    colnames(met_five)[x] = na.omit(conv[conv$name_mesa == colnames(met_five)[x], ]$metabid)
  }
}

length(colnames(met_one)[grep("M0", colnames(met_one))]) 
length(colnames(met_one)) 


for (x in 1:ncol(met_one)){
  met_one[,x] = as.numeric(met_one[,x])
} 
met_one = as.matrix(met_one)
met_one = log(met_one + 1)

for (x in 1:ncol(met_five)){
  met_five[,x] = as.numeric(met_five[,x])
} 
met_five = as.matrix(met_five)
met_five = log(met_five + 1)


```


# Calculate MetRS

## Exam 1
```{r}
#emphysema
file = "March/Results/models_copdgene/model_metabolite_pctEmph_Thirona_P2.rds"
  mod = readRDS(file)
  test = model_select(mod)
  test = replicate(met_one, mod$modlist[[test[1]]], lambda = test[2]) #
  test = merge(test, pd, by.x = "row.names", by.y = "sidno")
  test2 = lm(log(test$pctemph) ~ test$score + test$age1c + as.factor(test$race1c) + test$gender1 + as.factor(test$cig1c) + test$pkyrs1c + test$model)
  summary(test2)[[4]][2,4]  
  test = test %>% dplyr::select(Row.names, score)
  file = gsub("March/Results/models_copdgene/", "", file)
  file = gsub(".rds", "", file)
  names(test)[2] = file
  saveRDS(test, paste0("March/Results/predictions_mesa/", file, "_prediction_Exam1omics.rds"))  
  

```


## Exam 5
```{r}
#FEV1_P2  
file = "March/Results/models_copdgene/model_metabolite_FEV1_post_P2.rds"
  mod = readRDS(file)
  test = model_select(mod)
  test = replicate(met_five, mod$modlist[[test[1]]], lambda = test[2]) #
  test = merge(test, pd, by.x = "row.names", by.y = "sidno")
  test2 = lm(test$fev1ab5 ~ test$score + test$age5c + as.factor(test$race1c) + test$gender1 + as.factor(test$cig5c) + test$pkyrs5c + test$htcm5)
  summary(test2)[[4]][2,4] 
  test = test %>% dplyr::select(Row.names, score)
  file = gsub("March/Results/models_copdgene/", "", file)
  file = gsub(".rds", "", file)
  names(test)[2] = file
  saveRDS(test, paste0("March/Results/predictions_mesa/", file, "_predictions_Exam5omics.rds"))  
  

#Fvc_P2  
file = "March/Results/models_copdgene/model_metabolite_FVC_post_P2.rds"
  mod = readRDS(file)
  test = model_select(mod)
  test = replicate(met_five, mod$modlist[[test[1]]], lambda = test[2]) #
  test = merge(test, pd, by.x = "row.names", by.y = "sidno")
  test2 = lm(test$fvcab5 ~ test$score + test$age5c + as.factor(test$race1c) + test$gender1 + as.factor(test$cig5c) + test$pkyrs5c + test$htcm5)
  summary(test2)[[4]][2,4] 
  test = test %>% dplyr::select(Row.names, score)
  file = gsub("March/Results/models_copdgene/", "", file)
  file = gsub(".rds", "", file)
  names(test)[2] = file
  saveRDS(test, paste0("March/Results/predictions_mesa/", file, "_prediction_Exam5omics.rds"))  
  

#Fev/fvc_P2  
file = "March/Results/models_copdgene/model_metabolite_FEV1_FVC_post_P2.rds"
  mod = readRDS(file)
  test = model_select(mod)
  test = replicate(met_five, mod$modlist[[test[1]]], lambda = test[2]) #
  test = merge(test, pd, by.x = "row.names", by.y = "sidno")
  test2 = lm(test$fevfvcab5 ~ test$score + test$age5c + as.factor(test$race1c) + test$gender1 + as.factor(test$cig5c) + test$pkyrs5c + test$htcm5)
  summary(test2)[[4]][2,4] 
  test = test %>% dplyr::select(Row.names, score)
  file = gsub("March/Results/models_copdgene/", "", file)
  file = gsub(".rds", "", file)
  names(test)[2] = file
  saveRDS(test, paste0("March/Results/predictions_mesa/", file, "_prediction_Exam5omics.rds"))  
  
#finalgold_p2
file = "March/Results/models_copdgene/model_metabolite_finalGold_P2.rds"
  mod = readRDS(file)
  test = model_select(mod)
  test = replicate(met_five, mod$modlist[[test[1]]], lambda = test[2]) #
  test = merge(test, pd, by.x = "row.names", by.y = "sidno")
  test2 = lm(test$cpstgab5 ~ test$score + test$age5c + as.factor(test$race1c) + test$gender1 + as.factor(test$cig5c) + test$pkyrs5c)
  summary(test2)[[4]][2,4] 
  test = test %>% dplyr::select(Row.names, score)
  file = gsub("March/Results/models_copdgene/", "", file)
  file = gsub(".rds", "", file)
  names(test)[2] = file
  saveRDS(test, paste0("March/Results/predictions_mesa/", file, "_prediction_Exam5omics.rds"))  
  
#copd_p2
  file = "March/Results/models_copdgene/model_metabolite_copd_P2.rds"
  mod = readRDS(file)
  test = model_select(mod)
  test = replicate(met_five, mod$modlist[[test[1]]], lambda = test[2]) #
  test = merge(test, pd, by.x = "row.names", by.y = "sidno")
  test2 = glm(as.factor(test$cpgoldab5) ~ test$score + test$age5c + as.factor(test$race1c) + test$gender1 + as.factor(test$cig5c) + test$pkyrs5c, family = "binomial")
  summary(test2)[[13]][2,4] 
  test = test %>% dplyr::select(Row.names, score)
  file = gsub("March/Results/models_copdgene/", "", file)
  file = gsub(".rds", "", file)
  names(test)[2] = file
  saveRDS(test, paste0("March/Results/predictions_mesa/", file, "_prediction_Exam5omics.rds"))

#wa_pct
file = "March/Results/models_copdgene/model_metabolite_WallAreaPct_seg_Thirona_P2.rds"
  mod = readRDS(file)
  test = model_select(mod)
  test = replicate(met_five, mod$modlist[[test[1]]], lambda = test[2]) #
  test = merge(test, pd, by.x = "row.names", by.y = "sidno")
  test2 = lm(test$wa_pct ~ test$score + test$age5c + as.factor(test$race1c) + test$gender1 + as.factor(test$cig5c) + test$pkyrs5c + test$model)
  summary(test2)[[4]][2,4]  
  test = test %>% dplyr::select(Row.names, score)
  file = gsub("March/Results/models_copdgene/", "", file)
  file = gsub(".rds", "", file)
  names(test)[2] = file
  saveRDS(test, paste0("March/Results/predictions_mesa/", file, "_prediction_Exam5omics.rds"))  

#f2575ab5
file = "March/Results/models_copdgene/model_metabolite_FEF2575_post_P2.rds"
  mod = readRDS(file)
  test = model_select(mod)
  test = replicate(met_five, mod$modlist[[test[1]]], lambda = test[2]) #
  test = merge(test, pd, by.x = "row.names", by.y = "sidno")
  test2 = lm(test$f2575ab5 ~ test$score + test$age5c + as.factor(test$race1c) + test$gender1 + as.factor(test$cig5c) + test$pkyrs5c + test$htcm5)
  summary(test2)[[4]][2,4] 
  test = test %>% dplyr::select(Row.names, score)
  file = gsub("March/Results/models_copdgene/", "", file)
  file = gsub(".rds", "", file)
  names(test)[2] = file
  saveRDS(test, paste0("March/Results/predictions_mesa/", file, "_prediction_Exam5omics.rds"))  

#pefab5
file = "March/Results/models_copdgene/model_metabolite_PEF_post_P2.rds"
  mod = readRDS(file)
  test = model_select(mod)
  test = replicate(met_five, mod$modlist[[test[1]]], lambda = test[2]) #
  test = merge(test, pd, by.x = "row.names", by.y = "sidno")
  test2 = lm(test$pefab5 ~ test$score + test$age5c + as.factor(test$race1c) + test$gender1 + as.factor(test$cig5c) + test$pkyrs5c + test$htcm5)
  summary(test2)[[4]][2,4] 
  test = test %>% dplyr::select(Row.names, score)
  file = gsub("March/Results/models_copdgene/", "", file)
  file = gsub(".rds", "", file)
  names(test)[2] = file
  saveRDS(test, paste0("March/Results/predictions_mesa/", file, "_prediction_Exam5omics.rds"))  

  
```

# Wrap-Up

```{r}
Sys.time() - start

sessionInfo()
```
